Clipping fucking works
It worked all the time, I just was too stupid to realize the
tetrahedrons only looked smaller because they get cut in half.

There is now a #define option to assign "random" to each face to help
debug these issues. This may also be useful to debug certain problematic
shapes.