Implement custom Wavefront parser.
The previous one used waay too much memory and even after optimizations
it was quite slow. The new parser is:

- **Much** shorter & simpler
- A little slower (~98s -> ~101s for Lucy)
- Uses much less memory (5.4G -> 2.6G!)

It is also much easier to improve further with e.g. a custom allocator,
though it is very likely the biggest bottleneck now is std::stod, and
that's hard to replace without external libraries :(